Fix something??? (from pmmp)

+++ src/main/java/cn/nukkit/network/protocol/AnimatePacket.java
@@ -19,7 +19,7 @@ public class AnimatePacket extends DataPacket {
     @Override
     public void encode() {
         this.reset();
-        this.putVarInt((byte) action);
+        this.putVarInt(action);
         this.putEntityId(eid);
     }
